+++ src/input/input_options.rs
@@ -10,9 +10,11 @@ bitflags! {
 		const RESIZE = 0b0000_0010;
 		/// Enable undo and redo input handling
 		const UNDO_REDO = 0b0000_0100;
+		/// Search start
+		const SEARCH_START = 0b0000_1000;
 		/// Search handling
-		const SEARCH = 0b0000_1000;
+		const SEARCH = 0b0001_1000;
 		/// Help input handling
-		const HELP = 0b0001_0000;
+		const HELP = 0b0010_0000;
 	}
 }